Moving packet registering to ClientMod and MainMod
